backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
psutil==5.9.6
GPUtil==1.4.0
pynvml==11.5.0
pyyaml==6.0.1
pandas==2.1.4
datasets==2.15.0
transformers==4.36.0
torch==2.1.1
axolotl==0.4.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 